Add blocking test, fix lint
