+++ src/hypertrace/agent/instrumentation/django/__init__.py
@@ -1,17 +1,28 @@
+'''Hypertrace django instrumentor module wrapper.''' # pylint: disable=R0401
 import logging
+import traceback
 
+from opentelemetry.instrumentation.django import DjangoInstrumentor
+from opentelemetry.trace import Span
+from django.core.exceptions import PermissionDenied  # pylint:disable=C0415
+
+from hypertrace.agent import constants
 from hypertrace.agent.filter.registry import Registry
 from hypertrace.agent.instrumentation import BaseInstrumentorWrapper
 
+
 logger = logging.getLogger(__name__)  # pylint: disable=C0103
 
+
 class DjangoInstrumentationWrapper(BaseInstrumentorWrapper):
-
+    """wrapped class around django instrumentation"""
     def instrument(self):
-        from opentelemetry.instrumentation.django import DjangoInstrumentor
-        DjangoInstrumentor().instrument(request_hook=self.request_hook, response_hook=self.response_hook)
+        """configure django instrumentor w hooks"""
+        DjangoInstrumentor().instrument(request_hook=self.request_hook,
+                                        response_hook=self.response_hook)
 
-    def request_hook(self, span, request):
+    def request_hook(self, span: Span, request):
+        """django request hook before request is processed by app"""
         try:
             headers = list(dict(request.headers).items())
             body = request.body
@@ -24,11 +35,28 @@ class DjangoInstrumentationWrapper(BaseInstrumentorWrapper):
                                                     body)
             if block_result:
                 logger.debug('should block evaluated to true, aborting with 403')
-        except:
-            pass
+                # for some reason the span isnt marked as ended when we raise here
+                # since at this point middleware chain will halt
+                # we will end it
+                span.end()
+                raise PermissionDenied
+        except PermissionDenied as permission_denied:
+            raise permission_denied
+        except Exception as err:  # pylint:disable=W0703
+            logger.debug(constants.INST_RUNTIME_EXCEPTION_MSSG,
+                         'django request hook',
+                         err,
+                         traceback.format_exc())
 
 
-    def response_hook(self, span, request, response):
-        headers = list(dict(response.headers).items())
-        body = response.content
-        self.generic_response_handler(headers, body, span)
+    def response_hook(self, span, _request, response):
+        """django response hook before response is written out"""
+        try:
+            headers = list(dict(response.headers).items())
+            body = response.content
+            self.generic_response_handler(headers, body, span)
+        except Exception as err:  # pylint:disable=W0703
+            logger.debug(constants.INST_RUNTIME_EXCEPTION_MSSG,
+                         'django response hook',
+                         err,
+                         traceback.format_exc())
