+++ tests/django/manage.py
