+++ tests/django/setup.cfg
