+++ tests/django/testapp/settings.py
