+++ tests/django/tox.ini
